# docs/knowledge-graph/10-openapi.yaml
openapi: 3.1.0
info:
  title: TP Risk Management SMS — Safety Knowledge Graph Platform API
  version: 0.10.0-draft
  description: >
    DRAFT — controlled design document, requires approval before implementation.
    Backend REST surface for the Safety Knowledge Graph + AI Risk Intelligence Platform.
    Postgres is the system of record (03-postgresql-schema.sql); every write here is
    mirrored into Neo4j by the Graph Sync Service (01-enterprise-knowledge-graph-specification.md §4).
    See docs/PLATFORM_ARCHITECTURE_V2.md and docs/knowledge-graph/*.md for the design
    documents this contract implements.
    v0.2.0: Design Baseline v1.1 amendment — Emergency Planning (ACR-002) + Competency
    Management (ACR-003, supersedes ACR-001), approved 2026-08-04. Additive only —
    no removed/changed paths or schemas (02-development-standards.md §2 MINOR bump rule).
    v0.3.0: ACR-004 (approved 2026-08-11) — Investigation (1:1, INVESTIGATED_AS),
    incident-hazards linking (N:N, REVEALS, Option A bare-reference-list per ACR-004 §14),
    and incident-scoped Evidence. Drafted against ADR-003 (Investigation and Action are
    independent siblings of Incident, not a chain). Additive only — no removed/changed
    paths or schemas.
servers:
  - url: https://api.{environment}.vrtp-sms.example.net/v1
    variables:
      environment:
        default: dev
        enum: [dev, staging, prod]
security:
  - bearerAuth: []
...
paths:
...
  /hazards:
    get:
      tags: [Hazards]
      summary: List hazards
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - in: query
          name: asset_id
          schema: { type: string, format: uuid }
        - in: query
          name: category_concept_id
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Paginated hazard list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Hazard' } }
                  total: { type: integer }
    post:
      tags: [Hazards]
      summary: Create a hazard
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/HazardInput' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Hazard' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /hazards/{id}:
    parameters: [ { $ref: '#/components/parameters/id' } ]
    get:
      tags: [Hazards]
      summary: Get a hazard
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Hazard' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Hazards]
      summary: Update a hazard
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/HazardInput' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Hazard' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /risks:
    get:
      tags: [Risks]
      summary: List risks
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - in: query
          name: hazard_id
          schema: { type: string, format: uuid }
        - in: query
          name: current_rating
          schema: { type: string, enum: [Extreme, High, Medium, Low] }
      responses:
        '200':
          description: Paginated risk list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Risk' } }
                  total: { type: integer }
    post:
      tags: [Risks]
      summary: Create a risk (rating derived server-side — 07-inference-rules-catalogue.md R1)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RiskInput' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Risk' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /risks/{id}:
    parameters: [ { $ref: '#/components/parameters/id' } ]
    get:
      tags: [Risks]
      summary: Get a risk
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Risk' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    patch:
      tags: [Risks]
      summary: Update a risk. If current_rating is Extreme/High, sfarp_justification is validated (R4 SFARP gate).
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/RiskInput' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Risk' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { description: SFARP gate rejected the justification (07-inference-rules-catalogue.md R4) }

  # ── Controls & Critical Controls ─────────────────────────────────────────────────
...
  /performance-standards/{id}/verification-activities:
    parameters: [ { $ref: '#/components/parameters/id' } ]
    get:
      tags: [Verification]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/VerificationActivity' } } } } } }
    post:
      tags: [Verification]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/VerificationActivityInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/VerificationActivity' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /incidents/{id}/hazards:
    parameters: [ { $ref: '#/components/parameters/id' } ]
    get:
      tags: [Incidents, Hazards]
      summary: List hazards this incident revealed (REVEALS, N:N — safety.incident_hazards)
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Hazard' } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    post:
      tags: [Incidents, Hazards]
      summary: >
        Link a hazard this incident revealed (ACR-004 Option A — bare reference,
        no IncidentHazard schema; safety.incident_hazards has no columns beyond
        the incident_id/hazard_id composite key)
      requestBody: { content: { application/json: { schema: { type: object, required: [hazard_id], properties: { hazard_id: { type: string, format: uuid } } } } } }
      responses:
        '201': { description: Linked }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /incidents/{id}/hazards/{hazardId}:
    parameters:
      - { $ref: '#/components/parameters/id' }
      - { name: hazardId, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      tags: [Incidents, Hazards]
      summary: Unlink a hazard from this incident
      responses:
        '204': { description: Unlinked }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "AuthN/AuthZ provider TO BE CONFIRMED (architecture §2 AuthZ component) — Azure AD / Entra ID is the natural default given Azure hosting (architecture §8) but not yet decided."
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
    offset:
      name: offset
      in: query
      schema: { type: integer, default: 0 }
  responses:
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ValidationError:
      description: Request failed validation
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    ConceptRef:
      type: object
      description: Reference to an ontology concept (03-postgresql-schema.sql ontology.concepts)
      properties:
        concept_id: { type: string, format: uuid, nullable: true }
        pref_label: { type: string, description: "Read-only, resolved server-side", nullable: true }
...
    Hazard:
      allOf:
        - $ref: '#/components/schemas/HazardInput'
        - type: object
          properties:
            id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    HazardInput:
      type: object
      required: [name, description]
      properties:
        asset_id: { type: string, format: uuid, nullable: true }
        name: { type: string }
        description: { type: string }
        exposure_pathway: { type: string, nullable: true }
        possible_consequence: { type: string, nullable: true }
        category: { $ref: '#/components/schemas/ConceptRef' }
        energy_source: { $ref: '#/components/schemas/ConceptRef' }
        date_identified: { type: string, format: date }
        owner_person_id: { type: string, format: uuid }
        is_adh: { type: boolean, default: false, description: "Amusement device hazard — 11-safety-case-demonstration-model.md §3" }
        device_boundary_id: { type: string, format: uuid, nullable: true }
    Risk:
      allOf:
        - $ref: '#/components/schemas/RiskInput'
        - type: object
          properties:
            id: { type: string, format: uuid }
            inherent_rating: { type: string, enum: [Extreme, High, Medium, Low], readOnly: true, description: "Derived — 07-inference-rules-catalogue.md R1", nullable: true }
            current_rating: { type: string, enum: [Extreme, High, Medium, Low], readOnly: true, nullable: true }
    RiskInput:
      type: object
      required: [hazard_id, description]
      properties:
        hazard_id: { type: string, format: uuid }
        description: { type: string }
        cause: { type: string, nullable: true }
        inherent_likelihood: { type: integer, minimum: 1, maximum: 5 }
        inherent_consequence: { type: integer, minimum: 1, maximum: 5 }
        current_likelihood: { type: integer, minimum: 1, maximum: 5 }
        current_consequence: { type: integer, minimum: 1, maximum: 5 }
        target_likelihood: { type: integer, minimum: 1, maximum: 5 }
        target_consequence: { type: integer, minimum: 1, maximum: 5 }
        sfarp_justification: { type: string, description: "General WHS duty justification. For ADI/Chapter-9A-scoped risks, use serious_risk_justification and SFAIRP terminology in any generated Demonstration text — 09-regulatory-knowledge-model.md §3", nullable: true }
        status: { type: string, enum: [Open, Under Review, Closed] }
        review_date: { type: string, format: date }
        is_serious_risk: { type: boolean, default: false, description: "Chapter 9A serious risk classification (WHS Reg s.608B(1)(b)) — operator-defined threshold, see 09-regulatory-knowledge-model.md §6" }
        serious_risk_justification: { type: string, nullable: true }
...
    VerificationActivity:
      allOf:
        - $ref: '#/components/schemas/VerificationActivityInput'
        - type: object
          properties:
            id: { type: string, format: uuid }
            overdue: { type: boolean, readOnly: true, description: "07-inference-rules-catalogue.md R3" }
    VerificationActivityInput:
      type: object
      properties:
        method: { $ref: '#/components/schemas/ConceptRef' }
        frequency: { type: string, enum: [Daily, Weekly, Fortnightly, Monthly, Quarterly, Annual, Biennial, Other], nullable: true }
        due_date: { type: string, format: date }
        last_completed: { type: string, format: date }
        performed_by_person_id: { type: string, format: uuid }
        result: { type: string, nullable: true }
...